Parser re-org
