--- crates/winnow-datetime/src/sequence.rs
@@ -1,1 +1,0 @@
-
