+++ crates/winnow-iso8601/tests/formats/fractional_duration.rs
@@ -1,5 +1,6 @@
 use std::process::ExitCode;
 use winnow_datetime_assert::{define_format_tests, FractionalDurationAssertion};
+use winnow_iso8601::fractional_duration::fractional_duration;
 
 use winnow_datetime_assert::FormatAssertionBuilder;
 use winnow_datetime_assert::FormatCoverageBuilder;
