+++ crates/winnow-rfc3339/src/lib.rs
@@ -32,8 +32,6 @@ extern crate std;
 #[macro_use]
 extern crate alloc;
 
-pub mod parsers;
-
 /// date mod
 pub mod date;
 pub use date::parse_date;
