+++ crates/winnow-rfc3339/tests/formats/offset.rs
@@ -1,5 +1,6 @@
 use std::process::ExitCode;
 use winnow_datetime_assert::{define_format_tests, OffsetAssertion};
+use winnow_rfc3339::offset::offset;
 
 use winnow_datetime_assert::FormatAssertionBuilder;
 use winnow_datetime_assert::FormatCoverageBuilder;
